+++ tests/Fixtures/AppBundle/DataTable/Type/CustomQueryTableType.php
@@ -18,7 +18,7 @@ use Doctrine\ORM\QueryBuilder;
 use Omines\DataTablesBundle\Adapter\Doctrine\ORMAdapter;
 use Omines\DataTablesBundle\Column\TextColumn;
 use Omines\DataTablesBundle\DataTable;
-use Omines\DataTablesBundle\DataTableTypeInterface;
+use Omines\DataTablesBundle\DataTableType;
 use Tests\Fixtures\AppBundle\Entity\Employee;
 use Tests\Fixtures\AppBundle\Entity\Person;
 
@@ -27,7 +27,7 @@ use Tests\Fixtures\AppBundle\Entity\Person;
  *
  * @author Niels Keurentjes <niels.keurentjes@omines.com>
  */
-class CustomQueryTableType implements DataTableTypeInterface
+class CustomQueryTableType extends DataTableType
 {
     public function configure(DataTable $dataTable, array $options): void
     {