+++ CHANGELOG.md
@@ -9,6 +9,11 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - Updated the "per filter" twig files 
 - Updated datatable_html.html.twig to generate tfoot/thead filters for per column filtering using set html/js files
 
+### Breaking
+ - Added DataTableType Abstract class, you can extend this instead of implementing DataTableTypeInterface to maintain BC with anything existing
+ - Changed DataTableTypeInterface to support resolver options for both the type itself, and for the datatable to encapsulate settings in a type
+
+
 ## [0.8.0] - 2023-12-05
 ### Breaking
  - All interfaces and classes are now fully typed, this will likely require modifications in any custom extensions